update(s3_utils):update s3 saving functioning
